refactored code to make it working
